Extract body template renderer
